requirements.txt
fastapi
uvicorn[standard]
requests
aiohttp
python-dotenv
pydantic_settings